fix : Confirm Modal - isInRelay style
